+++ firmware/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Article name prefix "Getränk(e)" is now trimmed everywhere instead just in article selection
 - Added more user greetings
 - Added configuration to set Wifi country
 - Updated code to Rust Edition 2024, replaced JSON and font rendering libraries